dashboard/pyproject.toml
[tool.poetry]
name = "scrape_and_dash-dashboard"
version = "0.1.0"
description = "An e-commerce product analytics dashboard app that displays product data scraped and matched from multiple websites."
authors = ["Justin Andronicos"]

[tool.poetry.dependencies]
python = "^3.9"
dash = "^2.1.0"
dash-bootstrap-components = "^1.0.2"
Flask-Login = "0.6.1"
numpy = "^1.22.2"
pandas = "^1.4.0"
psycopg2 = "^2.9.3"
PyYAML = "^6.0"
SQLAlchemy = "^1.4.31"
Werkzeug = "^2.0.2"
models-items = "^0.1.0"

[tool.poetry.dev-dependencies]
alembic = "^1.7.6"
bandit = "^1.7.2"
black = "^22.1.0"
flake8 = "^4.0.1"
mypy = "^0.931"
pytest = "^7.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
